split home page into sections in separate files, make it responsive
