+++ apps/frontend/src/pages/home/home.tsx
@@ -1,0 +1,15 @@
+import HeroSection from './hero-section';
+import ProgramsSection from './programs-section';
+import CalendarSection from './calendar-section';
+import VideoSection from './video-section';
+
+export default function Home() {
+  return (
+    <>
+      <HeroSection />
+      <ProgramsSection />
+      <CalendarSection />
+      <VideoSection />
+    </>
+  );
+}
